feat: (#3) storybook 설정에 GlobalStyle(전역 스타일) 추가
